requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
qdrant-client==1.6.4
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
openai==1.65.1
pdf2image==1.17.0
pypdf==5.3.0
python-dotenv==1.0.0
requests==2.32.3
pandas==2.2.3
scikit-learn==1.6.1
tiktoken==0.9.0
sentence-transformers==3.4.1
rank_bm25
ragas
langchain-huggingface==0.1.2
huggingface-hub==0.29.1
